+++ app/src/main/java/com/example/customerapp/SettingFragment.java
@@ -81,8 +81,12 @@ public class SettingFragment extends Fragment {
 
     @SuppressLint("NotifyDataSetChanged")
     private void deleteAllQRandRecipients() {
-        addressBook.deleteAllRecipients(getContext());
-        Toast.makeText(getContext(), "Alle QR-Codes wurden gelöscht.", Toast.LENGTH_SHORT).show();
+        if (addressBook.getRecipients().isEmpty()) {
+            Toast.makeText(getContext(), "Es gibt keine QR-Codes zum Löschen.", Toast.LENGTH_SHORT).show();
+        } else {
+            addressBook.deleteAllRecipients(getContext());
+            Toast.makeText(getContext(), "Alle QR-Codes wurden gelöscht.", Toast.LENGTH_SHORT).show();
+        }
     }
 
 }
